+++ src/client/mock-server/data/golfers-list.ts
