+++ src/client/mock-server/index.ts
@@ -1,0 +1,2 @@
+export { worker } from './browser';
+// Don't export handlers through index. Crashes tests as it runs setUpWorker
