--- src/client/mocks/index.ts
@@ -1,2 +1,0 @@
-export { worker } from './browser';
-export { golfersStateData } from './data/golfers-state-data';
