+++ test/mocks/data/prize-money.ts
