+++ test/mocks/index.ts
@@ -1,1 +1,4 @@
 export { setupMockServer } from './helpers/mock-server';
+export { golfersStateData } from './data/golfers-state-data';
+export { teams } from './data/teams';
+export { prizeMoney } from './data/prize-money';
